refactor(fetch): file cleanup, add docstrings
